requirements.txt
fastapi
uvicorn
pillow
onnxruntime
numpy
opencv-python-headless
